fix: 1. merge from fxcore/evm 2. add intrarelayer contract

+++ app/cmd/update.go
@@ -4,6 +4,7 @@ import (
 	"bufio"
 	"errors"
 	"fmt"
+	"github.com/tendermint/tendermint/crypto/secp256k1"
 	"os"
 	"path/filepath"
 
@@ -24,13 +25,14 @@ const (
 	flagUpdateNodeKey = "update-node-key"
 	flagUpdatePrivKey = "update-priv-key"
 	flagUnsafe        = "unsafe"
+	flagKeyType       = "key-type"
 )
 
 func UpdateValidatorKeyCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "update-validator [secret]",
 		Short: "update validator node consensus private key file (.fxcore/config/priv_validator_key.json)",
-		Args:  cobra.ExactArgs(1),
+		Args:  cobra.RangeArgs(0, 1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			serverCtx := server.GetServerContextFromCmd(cmd)
 			rootDir := serverCtx.Viper.GetString(flags.FlagHome)
@@ -51,7 +53,15 @@ func UpdateValidatorKeyCmd() *cobra.Command {
 				}
 			}
 
-			secret := args[0]
+			secret := app.NewMnemonic()
+			if len(args) > 0 {
+				secret = args[0]
+			} else if !unsafe || !updateKey {
+				return fmt.Errorf("invalid params")
+			}
+			if len(secret) < 32 {
+				return fmt.Errorf("secret contains less than 32 characters")
+			}
 			pvKeyFile := serverCtx.Config.PrivValidatorKeyFile()
 			_ = os.Remove(pvKeyFile)
 			if err := tmos.EnsureDir(filepath.Dir(pvKeyFile), 0777); err != nil {
@@ -61,13 +71,21 @@ func UpdateValidatorKeyCmd() *cobra.Command {
 			if err := tmos.EnsureDir(filepath.Dir(pvStateFile), 0777); err != nil {
 				return err
 			}
-			valPrivKey := privval.NewFilePV(ed25519.GenPrivKeyFromSecret([]byte(secret)), pvKeyFile, pvStateFile)
-			valPrivKey.Save()
+			keyType := serverCtx.Viper.GetString(flagKeyType)
+			if keyType == "ed25519" {
+				valPrivKey := privval.NewFilePV(ed25519.GenPrivKeyFromSecret([]byte(secret)), pvKeyFile, pvStateFile)
+				valPrivKey.Save()
+			} else if keyType == "secp256k1" {
+				pk := secp256k1.GenPrivKeySecp256k1([]byte(secret))
+				valPrivKey := privval.NewFilePV(pk, pvKeyFile, pvStateFile)
+				valPrivKey.Save()
+			}
 			return nil
 		},
 	}
 	cmd.Flags().Bool(flagUpdatePrivKey, false, "Update ed25519 private key. Requires --unsafe.")
 	cmd.Flags().Bool(flagUnsafe, false, "Enable unsafe operations. This flag must be switched on along with all unsafe operation-specific options.")
+	cmd.Flags().String(flagKeyType, "ed25519", "Private key type, ed25519 or secp256k1.")
 	return cmd
 }
 
@@ -98,6 +116,11 @@ func UpdateNodeKeyCmd() *cobra.Command {
 			secret := app.NewMnemonic()
 			if len(args) > 0 {
 				secret = args[0]
+			} else if !unsafe || !updateKey {
+				return fmt.Errorf("invalid params")
+			}
+			if len(secret) < 32 {
+				return fmt.Errorf("secret contains less than 32 characters")
 			}
 			nodeKeyFile := serverCtx.Config.NodeKeyFile()
 			_ = os.Remove(nodeKeyFile)
